PUSH -> Added plugins cron function
