+++ .vscode/settings.json
@@ -11,6 +11,7 @@
         "Bethropolis",
         "ccaches",
         "configpath",
+        "cronfile",
         "dbconfigure",
         "doesinfoexist",
         "ealready",
