+++ storage/addons/Example/crons/job1.php
@@ -1,0 +1,5 @@
+<?php
+
+use MythicalSystemsFramework\Cli\Kernel;
+
+echo Kernel::translateColorsCode('Hello, World!&o');
